Clean up some duplicated code and improve comments.
No functional changes.
